Added Object Generation Outline
There is now the WorldGenerator class which stores all of the generator scripts. These generator scripts could be for terrain, clouds, weather effects, or even time of day.
