+++ unity/easyWorld/Assets/Scenes/Patrick-Terrain/GeminiTerrain.unity
@@ -1059,7 +1059,53 @@ MonoBehaviour:
   serverHandler: {fileID: 2088404244}
   resetTextOnButtonClick: 0
   inputField: {fileID: 1956810249}
-  terrainGenerator: {fileID: 70678443}
+  worldGenerator: {fileID: 1364701231}
+--- !u!1 &1364701230
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1364701232}
+  - component: {fileID: 1364701231}
+  m_Layer: 0
+  m_Name: WorldGenerator
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1364701231
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1364701230}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 981ee8a2cc0511e41ac2772ded93a20a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  generators:
+  - {fileID: 70678443}
+--- !u!4 &1364701232
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1364701230}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 78.36102, y: 151.55367, z: -115.983444}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1956810247
 GameObject:
   m_ObjectHideFlags: 0
@@ -1443,4 +1489,5 @@ SceneRoots:
   - {fileID: 513057399}
   - {fileID: 2088404245}
   - {fileID: 1327592065}
+  - {fileID: 1364701232}
   - {fileID: 70678444}
