+++ unity/easyWorld/Assets/Scenes/Patrick-Terrain/Generator.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using UnityEngine;
+
+public abstract class Generator : MonoBehaviour
+{
+    public abstract IEnumerator Generate(WorldInfo worldInfo);
+
+    public abstract void Clear();
+}
